refactor: export options datatypes + skeleton of the config page
