+++ src/routes/config/+page.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+
+</script>
+
+<svelte:head>
+	<title>Configuration - Svelte Local GPX Viewer</title>
+</svelte:head>
+
+<h1>App Settings</h1>
